// src/components/toast/dependencies/toast.types.tsx
import type { ReactNode } from 'react';
//AUTO GENERATED BY DYVIX UI ENGINE - DO NOT MANUALLY EDIT.

export type DyvixToastAnimation =
  | 'fade'
  | 'bubble'
  | 'zoom'
  | 'unfold'
  | 'glitch'
  | 'pulse'
  | 'aurora'
  | 'drop'
  | 'flip'
  | 'glide'
  | 'drift'
  | 'float'
  | 'swing'
  | 'slideRight'
  | 'spiral'
  | 'bounce';

export type DyvixToastPosition =
  | 'top-left'
  | 'top-right'
  | 'bottom-left'
  | 'bottom-right'
  | 'top-center'
  | 'bottom-center';

export type DyvixToastType = 'Error' | 'Success' | 'Warning' | 'Info';

/*--!/--*/

// Edit here safely

export interface DyvixToastAnimationConfig {
  animation: DyvixToastAnimation | string;
  type: string;
  'default-duration': number;
  ease: string;
  from: Record<string, any>;
  to: Record<string, any>;
}

export interface DyvixToastContainerProps {
  position?: DyvixToastPosition;
  segments?: number;
  duration?: number;
  animation?: DyvixToastAnimation | null | undefined;
}
export interface ToastItem {
  id: string;
  message: string;
  type: string;
  [key: string]: any;
}

export interface DyvixToast {
  className?: string;
  message: string;
  animation?: DyvixToastAnimationConfig | null | undefined;
  type: DyvixToastType;
  onClose: () => void;
  duration?: number;
}